+++ CHANGELOG.md
@@ -6,10 +6,15 @@
 ### Features
 - Improved `Dockerfile` and added option to install additional packages during image build ([#28](https://github.com/AleksanderWWW/abcd-graph/pull/28))
 - Added a top-level `ABCDCommunity` class to represent user-facing community properties ([#31](https://github.com/AleksanderWWW/abcd-graph/pull/31))
+- Added a method to retrieve membership list for the graph ([#34](https://github.com/AleksanderWWW/abcd-graph/pull/34))
+- Added ground truth community property to exported graph attributes ([#34](https://github.com/AleksanderWWW/abcd-graph/pull/34))
+- Added outlier generation to the graph ([#36](https://github.com/AleksanderWWW/abcd-graph/pull/36))
 
 ### Breaking Changes
 - Renamed the top level graph object from `Graph` to `ABCDGraph` ([#31](https://github.com/AleksanderWWW/abcd-graph/pull/31))
 - Made retrieving edge and community lists the responsibility of the `ABCDGraph` object ([#31](https://github.com/AleksanderWWW/abcd-graph/pull/31))
+- Change `ABCDParams` arguments and get rid of pydantic ([#36](https://github.com/AleksanderWWW/abcd-graph/pull/36))
+- Include number of vertices and number of outliers to the params ([#36](https://github.com/AleksanderWWW/abcd-graph/pull/36))
 
 
 ## abcd-graph 0.2.1